// src/enums/DK64Barrel.ts
export enum DK64Barrel {
  BattyBarrelBandit = "Batty Barrel Bandit",
  BeaverBother = "Beaver Bother",
  BigBugBash = "Big Bug Bash",
  BusyBarrelBarrage = "Busy Barrel Barrage",
  KrazyKongKlamour = "Krazy Kong Klamour",
  KremlingKosh = "Kremling Kosh",
  MadMazeMaul = "Mad Maze Maul",
  MinecartMayhem = "Minecart Mayhem",
  PerilPathPanic = "Peril Path Panic",
  SearchlightSeek = "Searchlight Seek",
  SpeedySwingSortie = "Speedy Swing Sortie",
  SplishSplashSalvage = "Splish-Splash Salvage",
  StashSnatch = "Stash Snatch",
  StealthySnoop = "Stealthy Snoop",
  TeeteringTurtleTrouble = "Teetering Turtle Trouble"
}