// src/enums/DK64Category.ts
export enum DK64Category {
  GB = "Golden Banana",
  Blueprint = "Blueprint",
  Medal = "Banana Medal",
  HelmMedal = "Banana Medal (Hideout Helm)",
  Crown = "Battle Crown",
  Key = "Boss Key",
  Fairy = "Banana Fairy",
  CompanyCoin = "Company Coin",
  RainbowCoin = "Rainbow Coin",
  ColoredBanana = "Colored Banana",
  Warp = "Warp"
}